+++ src/home_robot/home_robot/motion/base.py
@@ -3,8 +3,8 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-from home_robot.agent.motion.robot import Robot
-from home_robot.agent.motion.space import Space
+from home_robot.motion.robot import Robot
+from home_robot.motion.space import Space
 
 """
 This just defines the standard interface for a motion planner
