+++ src/home_robot/home_robot/motion/linear.py
@@ -4,8 +4,8 @@
 # LICENSE file in the root directory of this source tree.
 import numpy as np
 
-from home_robot.agent.motion.base import Planner
-from home_robot.agent.motion.robot import STRETCH_STANDOFF_WITH_MARGIN, HelloStretchIdx
+from home_robot.motion.base import Planner
+from home_robot.motion.robot import STRETCH_STANDOFF_WITH_MARGIN, HelloStretchIdx
 
 
 class LinearPlanner(Planner):
